requirements.txt
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.8.0.76
Pillow==9.5.0
streamlit==1.25.0
tensorflow==2.17.1
